# openapi.yaml
openapi: 3.0.0
info:
  title: Logistic Backend API
  description: A comprehensive logistic management system API built with NestJS
  version: 1.0.0
  contact:
    name: Logistic Team
    url: https://logistic-backend.com
    email: support@logistic-backend.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.logistic-backend.com
    description: Production server
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Login successful
                data:
                  user:
                    id: uuid
                    email: user@example.com
                    firstName: John
                    lastName: Doe
                    role: user
                    status: active
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: false
                message: Invalid credentials
                errors:
                  - Invalid email or password
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: User registered successfully
                data:
                  user:
                    id: uuid
                    email: user@example.com
                    firstName: John
                    lastName: Doe
                    role: user
                    status: active
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: false
                message: Validation failed
                errors:
                  - Email already exists
  /api/v1/auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Get current authenticated user profile
      operationId: getProfile
      security:
        - JWT-auth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Profile retrieved successfully
                data:
                  id: uuid
                  email: user@example.com
                  firstName: John
                  lastName: Doe
                  role: user
                  status: active
                  phone: +1234567890
                  createdAt: '2023-01-01T00:00:00.000Z'
                  updatedAt: '2023-01-01T00:00:00.000Z'
        '401':
          description: Unauthorized - Invalid or missing token
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate a new access token using the current valid token
      operationId: refreshToken
      security:
        - JWT-auth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Token refreshed successfully
                data:
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Unauthorized - Invalid or expired token
...
components:
  securitySchemes:
    JWT-auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
    ApiResponseDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
          example: true
        message:
          type: string
          description: Response message
          example: Operation completed successfully
        data:
          description: Response data payload
        meta:
          type: object
          description: Pagination metadata
          properties:
            page:
              type: number
              example: 1
            limit:
              type: number
              example: 10
            total:
              type: number
              example: 100
            totalPages:
              type: number
              example: 10
        errors:
          type: array
          items:
            type: string
          description: Array of error messages
          example:
            - Validation failed
            - Invalid input
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          minLength: 6
          description: User password
          example: password123
    RegisterDto:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          minLength: 6
          description: User password
          example: password123
        firstName:
          type: string
          description: User first name
          example: John
        lastName:
          type: string
          description: User last name
          example: Doe
        phone:
          type: string
          description: User phone number
          example: +1234567890
...